Update building-running.md
I think the order of the env var and script call is inverted

Also PostGreSQL also needs to be set up properly with the right users, so these instructions really need to be revised to state that.  It would be very helpful if there
were cabal instructions as well as nix ones (the installation is broken, and there is no target for the build).  And the binary cache link is wrong.