playChainFragment: run a chain-sync client against a server
This patch implements
```
playChainFragment :: Config blk
                  -> ChainActions
                  -> ChainSyncClient blk (Point blk) (Tip blk) m a
                  -> m a
```
It runs a chain sync client against a server without using network
connection, and thus no serialisation of data is used.  This could speed
up execution of tests.

The generator of 'ChainActions' does not guarantee that the rollbacks
are on the chain.  This should be implemented as the 'serverDriver'
relies on this property.

This patch provides an 'Arbitrary' instance function for 'ChainActions',
but they are not tested, and thus their behaviour / distribution is not
guaranteed to be the right one.